Refactored the AssetPlayerBox into smaller components
